pyproject.toml
[project]
name = "stress-temperature"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "cartopy>=0.23.0",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scipy>=1.13.1",
    "st-pages>=1.0.1",
    "streamlit>=1.49.1",
]
